requirements.txt
numpy==1.22.3
pandas==1.4.2
python-dotenv==0.20.0
requests==2.27.1
ta==0.10.1
urllib3==1.26.9
matplotlib==3.5.2
mplfinance==0.12.9b0
ccxt==1.84.78
typer==0.4.1
python-binance==1.0.16